refactor: move docstatus to a separate file
